Monse Sandoval-Malherbe
I chose any main series Pokemon game for the game I used in this assignment.  For the first condition, I was imagining the battle taking place was a fight versus your Charmander and your rivals Squirtle. In the second condition, MC just refers to the playable main character. The third condition, I imagined you were in  a battle with a member from the Elite Four.